pyproject.toml
[project]
name = "omniservice-ai"
version = "0.1.0"
description = "Agentic System for Service AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "cohere>=5.20.4",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "langchain-anthropic>=1.3.3",
    "langchain-core>=1.2.12",
    "langgraph>=1.0.8",
    "neo4j>=6.1.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
    "weaviate-client>=4.19.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
